+++ bootstrap/cache/packages.php
@@ -31,6 +31,33 @@
       0 => 'Livewire\\LivewireServiceProvider',
     ),
   ),
+  'nativephp/desktop' => 
+  array (
+    'aliases' => 
+    array (
+      'Dock' => 'Native\\Desktop\\Facades\\Dock',
+      'Menu' => 'Native\\Desktop\\Facades\\Menu',
+      'Shell' => 'Native\\Desktop\\Facades\\Shell',
+      'Screen' => 'Native\\Desktop\\Facades\\Screen',
+      'System' => 'Native\\Desktop\\Facades\\System',
+      'Window' => 'Native\\Desktop\\Facades\\Window',
+      'MenuBar' => 'Native\\Desktop\\Facades\\MenuBar',
+      'Process' => 'Native\\Desktop\\Facades\\Process',
+      'Updater' => 'Native\\Electron\\Facades\\Updater',
+      'Settings' => 'Native\\Desktop\\Facades\\Settings',
+      'Clipboard' => 'Native\\Desktop\\Facades\\Clipboard',
+      'ContextMenu' => 'Native\\Desktop\\Facades\\ContextMenu',
+      'QueueWorker' => 'Native\\Desktop\\Facades\\QueueWorker',
+      'ChildProcess' => 'Native\\Desktop\\Facades\\ChildProcess',
+      'Notification' => 'Native\\Desktop\\Facades\\Notification',
+      'PowerMonitor' => 'Native\\Desktop\\Facades\\PowerMonitor',
+      'GlobalShortcut' => 'Native\\Desktop\\Facades\\GlobalShortcut',
+    ),
+    'providers' => 
+    array (
+      0 => 'Native\\Desktop\\NativeServiceProvider',
+    ),
+  ),
   'nesbot/carbon' => 
   array (
     'providers' =>